modify user

+++ src/main/java/com/blog/yongyu/demo/Constroller/UserController.java
@@ -14,6 +14,7 @@ import com.blog.yongyu.demo.Service.UserInfoService;
 import com.blog.yongyu.demo.Service.UserRoleService;
 import com.blog.yongyu.demo.Utils.ResultUtils;
 import javafx.geometry.Pos;
+import net.minidev.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -54,8 +55,12 @@ public class UserController {
         return ResultUtils.error(res, msg[res]);
     }
 
-    @RequestMapping("/deleteUserInfo")
+    @RequestMapping(value = "/deleteUserInfo",method = RequestMethod.POST)
     public DataResult deleteUserInfo(@RequestParam("uid") Long uid) {
+        LoginInfor logiInfo = loginInfoService.getLogiInfo();
+        if (logiInfo == null ||!Objects.equals(logiInfo.getRoleId(), BaseRole.AdminId)) {
+            return ResultUtils.error(2, "没有权限");
+        }
         Integer res = userInfoService.Delete(uid);
         if (res == 0) {
             return ResultUtils.success();
@@ -76,28 +81,44 @@ public class UserController {
     public DataResult findAll() {
         List<UserRole> allUserRole = userRoleService.findAll();
         Map<String, Integer> map = new HashMap<>();
+        List<JSONObject> list = new ArrayList<>();
         int i=0;
         for (UserRole ur : allUserRole) {
+            JSONObject jsonObject = new JSONObject();
             if (map.containsKey(ur.getAccount())) {
                 if (ur.getIsDefault().equals(UserRole.ISDEFAULT.isDefault.toString())) {
                     allUserRole.set(map.get(ur.getAccount()), ur);
+                    jsonObject.put("account", ur.getAccount());
+                    jsonObject.put("userName", ur.getUserInfo().getUserName());
+                    jsonObject.put("sex", ur.getUserInfo().getSex());
+                    jsonObject.put("email", ur.getUserInfo().getEmail());
+                    jsonObject.put("phone", ur.getUserInfo().getPhone());
+                    jsonObject.put("status", ur.getUserInfo().getStatus());
+                    jsonObject.put("roleName", ur.getRole().getRoleName());
                 }
-                allUserRole.remove(i);
             } else {
                 map.put(ur.getAccount(), i++);
+                jsonObject.put("account", ur.getAccount());
+                jsonObject.put("userName", ur.getUserInfo().getUserName());
+                jsonObject.put("sex", ur.getUserInfo().getSex());
+                jsonObject.put("email", ur.getUserInfo().getEmail());
+                jsonObject.put("phone", ur.getUserInfo().getPhone());
+                jsonObject.put("UserStatus", ur.getUserInfo().getStatus());
+                jsonObject.put("roleName", ur.getRole().getRoleName());
             }
+            list.add(jsonObject);
         }
 //        List<UserInfo> all = userInfoService.findAll();
-        return ResultUtils.success(allUserRole);
+        return ResultUtils.success(list,list.size());
     }
     @RequestMapping("/resetPwd")
     public DataResult resetPwd(@RequestParam("id")Long id){
-        Optional<UserInfo> userById = userInfoService.findUserById(id);
-        if (!userById.isPresent()) {
+        UserInfo userById = userInfoService.findUserById(id);
+        if (userById==null) {
             return ResultUtils.error(1, "用户不存在");
         }
-        userById.get().setInitPassword();
-        userInfoService.modify(userById.get());
+        userById.setInitPassword();
+        userInfoService.modify(userById);
         return ResultUtils.success();
     }
